+++ libs/common/src/auth/models/request/update-temp-password.request.ts
@@ -1,4 +1,4 @@
-import { OrganizationUserResetPasswordRequest } from "../../abstractions/organization-user/requests";
+import { OrganizationUserResetPasswordRequest } from "../../../abstractions/organization-user/requests";
 
 export class UpdateTempPasswordRequest extends OrganizationUserResetPasswordRequest {
   masterPasswordHint: string;
